requirements.txt
streamlit==1.28.0
pandas==2.1.0
plotly==5.17.0
gspread==5.12.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.104.0
numpy==1.24.0
Flask==2.3.3
flask-cors==4.0.0